moved addClub func to controllers
